% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_timeout.R
\name{measure_all_timeout}
\alias{measure_all_timeout}
\alias{measure_timeout}
\title{Measure timeout of log}
\usage{
measure_all_timeout(log_df)

measure_timeout(log_df, N = 2L)
}
\arguments{
\item{log_df}{A \code{data.frame} that records the log.}

\item{N}{Threshold of continuous errors to be considered as fatal.}
}
\value{
A data frame with following columns:
\itemize{
\item address: (chr) IPv4 addresses that suffered timeout.
\item start: (dttm) When server started to timeout.
\item end: (dttm) When server resumed or \code{NA} if not resumed.
\item n_timeout: (int) Number of timeout before resume.
}

data.frame
}
\description{
\code{measure_timeout()} measures timeout errors that continuously occurs \code{N}
times or more. \code{N = 1} is equivalent to \code{measure_all_timeout()}.
}
\examples{
log_df <- data.frame(
  timestamp = lubridate::ymd_hms(20200101000000 + seq(5)),
  address = "192.168.1.1/24",
  ping = c(1, NA, 1, NA, NA)
)

measure_timeout(log_df, N = 1L) # equivalent to measure_all_timeout(log)
measure_timeout(log_df, N = 2L)

}
