% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_overload.R
\name{measure_overload}
\alias{measure_overload}
\title{Measure overload}
\usage{
measure_overload(log_df, m = 2L, t = 10, N = 1L)
}
\arguments{
\item{log_df}{A \code{data.frame} that records the log.}

\item{m}{A window size of rolling mean of ping time.}

\item{t}{Upper limit of rolling mean of ping time.}

\item{N}{Threshold of continuous errors to be considered as fatal.}
}
\value{
A data frame with following columns:
\itemize{
\item address: (chr) IPv4 addresses that suffered timeout.
\item start: (dttm) When server started to timeout.
\item end: (dttm) When server resumed or \code{NA} if not resumed.
\item n_timeout: (int) Number of timeout before resume.
}

data.frame
}
\description{
Summarize the overload conditions when rolling mean of ping exceeds the
limit. If a window contains timeout, it is not considered as overload
regardless of any finite values within the window.
Note that no-response is discarded instead of being treated as timeout by
\code{measure_timeout()}.
}
\examples{
log_df <- data.frame(
  timestamp = lubridate::ymd_hms(20200101000000 + seq(6)),
  address = "192.168.1.1/24",
  ping = c(1, 10, NA, 10, 10, NA)
)
measure_overload(log_df, m = 1L, t = 10)
measure_overload(log_df, m = 2L, t = 10)

}
