% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_log.R
\name{read_log}
\alias{read_log}
\title{Read and parse log file as a data frame.}
\usage{
read_log(file)
}
\arguments{
\item{file}{A path to the CSV file or a CSV formated string.
First row starts with values without headers.}
}
\value{
A data frame object with three columns sorted by these columns:
\itemize{
\item timestamp: (dttm) Ordered timestamps of ping.
\item address: (chr) IPv4 address of server.
\item ping: (dbl) Elapsed time in msec, or \code{NA} in case of timeout.
}
}
\description{
Read and parse log file as a data frame.
}
\examples{
# Prepare data
log_csv <- paste(
  "20200101000001, 192.168.1.1/24, 5",
  "20200101000002, 192.168.1.1/24, 8",
  sep = "\n"
)
path <- tempfile()
writeLines(log_csv, path)

# Read from path and string.
read_log(path)
read_log(I(log_csv))


}
